# Optimize the code for retrieving annotations