--- a/changelog.d/20240422_133304_maria_optimize_retrieving_annotations.md
+++ b/changelog.d/20240422_133304_maria_optimize_retrieving_annotations.md
@@ -1,0 +1,4 @@
+### Fixed
+
+- Optimized memory usage when retrieving annotations by disabling internal Django QuerySet caching
+  (<https://github.com/cvat-ai/cvat/pull/7748>)