@Marishka17 2024-04-18
@azhavoro, @zhiltsov-max 
I've re-measured metrics again. Besides that, I figured out that the `prefetche_related` used here was useless since Django implementation. To put it briefly, `prefetch_related(...)` will not work if `values(...)` is used because the cache is saved to `_prefetched_objects_cache` object attribute, but after using `values(...)` `ValuesIterable` returns a queryset with `dicts` and it is not possible to set an attribute to dict. ([Link to the code block](https://github.com/django/django/blob/2be37b253341cfd1f1363c533e6f896230f047a7/django/db/models/query.py#L2326-L2342))

  
| Queryset | Time for retreiving annotations (min) | Number of Queries | Memory usage (GB) | Chunk size |
| ------ | ------ | ------ | ------ | ------ |
| original | 17.8 | 17 | 18.3 | default 100 |
| original + iterator | 18.3 | 17 | 5.6 | default 2000 |
| original - prefetch + iterator | 18 | 17 | 5.6 | default 2000 |
| original - prefetch + iterator | 18.5 | 17 | 5.6 | 10000 |